refactor: extract callable handling into Callables trait
